added Readme. Rename spaced to aet to avoid collision on Pypi
