+++ aether-utils/aether-consumer/tests/test_sdk.py
@@ -19,7 +19,7 @@
 # under the License.
 
 from . import *
-from aether.consumer import KafkaConsumer
+from aet.consumer import KafkaConsumer
 
 # Test Suite contains both unit and integration tests
 # Unit tests can be run on their own from the root directory
